refactor(app): consolidate arena stores, hooks, contexts, and lib into feature
