+++ src/features/app/components/app-home.tsx
@@ -3,7 +3,7 @@ import { AppArenaSideEffects } from './app-arena-side-effects';
 import { AppHUD } from './hud';
 import { ResultDialog } from './match-result';
 import { Scoreboard } from './scoreboard';
-import { useMatchResult } from '@/hooks/use-match-result';
+import { useMatchResult } from '@/features/app/hooks/use-match-result';
 import { useFinishMatch } from '@/features/app/hooks/use-finish-match';
 import { Navbar } from '@/components/navigation/navbar';
 import { Dialog } from '@/components/ui/dialog';
