+++ src/features/app/components/settings/tabs/standard-settings.tsx
@@ -3,7 +3,7 @@ import { Plus } from 'lucide-react';
 import { PlayerAvatar } from '../../hud/player-avatar';
 import { avatarGroup, playerGroup } from '../constant/form';
 import { CommonSettings } from './common-settings';
-import { useSettings } from '@/contexts/settings';
+import { useSettings } from '@/features/app/contexts/settings';
 import { useAppForm } from '@/components/form/hooks';
 import { StandardSettingsSchema } from '@/lib/validations';
 import {
