+++ src/features/app/hooks/use-advance-settings.tsx
@@ -4,7 +4,7 @@ import { getTournamentFields } from '../components/settings/constant/form';
 import { resolveGroupComboboxStatus } from '../components/settings/lib/group';
 import { resolveMatchComboboxRow } from '../components/settings/lib/match';
 import type { QueryClient, UseQueryResult } from '@tanstack/react-query';
-import type { AdvanceFormData } from '@/contexts/settings/context';
+import type { AdvanceFormData } from '@/features/app/contexts/settings/context';
 import type {
   SelectionGroupRow,
   SelectionMatchRow,
