+++ src/features/app/lib/hud/health-utils.ts
