+++ src/features/app/lib/scoreboard/hit-types.ts
